[data-poll-manager] no parent, no data poll!
The goal of this change is to ensure that data polls (IEEE 802.15.4
Data Request commands) are not sent when there is no parent (or parent
candidate) or the parent is not in valid/restoring states.

In both `DataPollManager` and `MeshForwarder` classes, new checks are
added to ensure that there is a valid parent before enqueuing/sending
a data poll message. Also in `MeshForwarder`, the MAC source and
destination addresses for a data poll frame are determined when
preparing the data poll message (in `GetDirectTransmission()`) instead
of from `SendPoll()` call.

This commit also adds two helper methods in `Mac::Frame` class for the
common operation of setting source/destination addresses on a frame using
a `Mac::Address` as input (which may contain either a short or an
extended address).