[ncp] handle `NOT_FOUND` and `ALREADY` errors during `VALUE_INSERT`/`VALUE_REMOVE` commands (#2103)
This commit changes the error type returned from `VALUE_IS` and
`VALUE_REMOVE` spinel commands.  If a `VALUE_REMOVE` is issued where the
item being removed is not in the list, instead of passing up `NOT_FOUND`
error, a `LAST_STATUS` with `STATUS_OK` is used. Similarly the behavior
is changed for the `VALUE_INSERT` operation and error status `ALREADY`.

This change helps ensure driver/user considers the operation as a
success while still being able to distinguish (if required) whether
the value was actually removed by checking if we get a `VALUE_REMOVED`
response versus `LAST_STATUS(OK)` response.