[meshcop] Give priority to better matches during join, not channel
Previous behaviour of the joiner would result in the joiner joining
the network with the highest channel number, rather than the most
suitable. This commit utilises the priority specified in thread spec
1.1.1 section 8.4.4.1.2 to choose the most suited network for joining.
This removes the bug where it would be impossible to join the desired
network due to another network on a higher channel allowing all joiners
in the steering data.

There is still further work to be done to fully match the spec here, as
currently there is no queue mechanic to try a different network if the
DTLS handshake fails.