[mle] fix partition merging bug when determining singleton partitions (#2203)
The Thread 1.1.1 Specification currently defines a "singleton" as a Thread
network partition that consists of a single router that has no
router-eligible end device children.  At the same time, Thread 1.1.1 does
not provide a way for a router to decipher between a Full End Device (FED)
and a router-eligible end device (REED).  This ambiguity can cause two
partitions to avoid merging, since each partition may determine that they
are higher priority than the other.

This commit makes the change to counting only active routers, which
eliminates the ambiguity above.