[mle] do not restore last partition id after boot
While the previous partition ID is stored in non-volatile memory, the last
known router ID sequence for that previous partition is not.  As a result,
a node coming out of reset that initially fails to reattach to the previous
partition may not be able to reattach until the router ID sequence value
wraps.  This commit also limits the time window with which to avoid the
previous partition ID.