[mle] suppress Data Response when trying to update network data (#2335)
There may be short windows where a device does not have the proper Thread
Network Data.  For example, when a device is coming out of reset and
resynchronizing with the rest of the network.

This commit suppresses MLE Data Response transmissions while the device is
trying to receive the latest Thread Network Data from a neighbor.  This
eliminates the possibility that the device may propagate inconsistent
Thread Network Data.

This commit fixes spurious failures in Cert_9_2_15_PendingPartition.py.