[radio] change radio API to support more advanced transmit power configs
The transmit power configuration in the existing radio API (i.e.
`otPlatRadioSetDefaultTxPower()` abstracted the radio power as an int8_t
in units of dBm.  This was overly constraining for platforms that had more
advanced ways of configuring the transmit power.

At the same time, Thread (and OpenThread), does not employ any form of
transmit power control.  As a result, while OpenThread provides APIs to
control the transmit power, it simply passes the transmit power value as
an opaque value.

This commit changes makes the following changes:

1. `otPlatRadioSetDefaultTxPower(int8_t)` has been changed to
   `otPlatRadioSetTransmitPower(uint32_t)`.

2. The transit power configuration increased to 32 bits, allowing for more
   advanced transmit power configurations.  The interpretation of this
   32-bit value is now radio-specific.

3. The `otRadioFrame` structure is now updated to separate the transmit and
   received power fields.  The transmit power field is now mTxPowerConfig.
   The received signal strengy field is now mRssi.