[NCP] Added support for external NCP vendor hook
This change allows developers to add a hook source file to the NCP
build that contains hook for implementing vendor-specific commands.
This is done via a new configure script argument,
`--with-ncp-vendor-hook-source=<C++FILE>`, with the argument being a
single C++ source file containing the hook.

This isn't a terribly elegant solution to this sort of problem, so
long-term we should consider major refactoring of the NCP codebase
to enable the flexibility of allowing the vendor to implement
a library that can be linked in, instead of a file which must be
compiled in.