[cli] make CLI buffer sizes configurable (#2351)
* [core] Make CLI buffer sizes configurable.

At the moment, we allocate 512 bytes for a receive buffer and 1kB for
transmit, which while it doesn't seem a lot, is in fact a big commitment
for a CLI interface.  In cases where the platform does its own
buffering, this can be significantly reduced.

These add definitions that mirror the current settings so they may be
changed.

* [cli] Define UART buffers according to config.

Use the definitions produced in the previous commit to set the buffer
sizes.

* [core, cli] Rename OPENTHREAD_CONFIG_CLI_LINE_LENGTH

to `OPENTHREAD_CONFIG_CLI_MAX_LINE_LENGTH` as per review in pull request 2351.