[tests] fix the unstable test case Cert_5_5_08_SplitRoutersLostLeader
The unstable failing lies in when Leader has lower partition id and switches to better partition
formed by original routers.  The default RouterSelectionJitter of Leader is 120s, however the
migration waiting time in script is 60s, thus there are chances that the ping happens when the
Leader just becomes router in the new partition but before link sychronization, causing the ping
packet dropped on ROUTER3.

Two options to resolve this: 1) increase the waiting time for migration to a value > 120s or
2) configure router selection jitter to a small value to relax the waiting time requirements.