[mle] do not clear router table state until role transitions to router
Prior to this commit, a REED attempting to become a router clears its
routing table state when sending the Address Solicit message.  This causes
the REED to improperly think that it is a singleton when determining whether
or not to merge with another partition.

This commit only updates the routing table state when the REED actually
transitions to becoming a router.