[mle] fix REED's failing to become router due to the invalid soliciting state
It would take some time for the Leader to establish the routing to the new router in linear
topology, which may happenly cause following FTD trapped in REED state because Leader
would drop the address reply as it thinks the REED's parent is not reachable in this
time window. And the REED would not retry any further as the flag mAddressSolicitPending is
only be cleared when receiving the address reply.

This commit allows further address solicit attempt under proper condition and restricts the interval
between two sucessive address solicit attempts as well.