[spi-hdlc-adapter] enhance the retry logic when slave is busy
This commit changes the `spi-hdlc-adapter` retry logic when the slave
is busy (it refuses a transmission). There are three levels, initially
the retries happen almost immediately, then up to certain number of
attempts at a fast rate and finally it drops to a slow rate.

This change helps with the average delay of spinel message exchange
between NCP and host (connected using an SPI interface).