[jam-detector] pause jam detection when thread is disabled
Once jamming detection is enabled/started, `JamDetector`  will also
monitor the Thread device role. RSSI sampling is stopped if the device
role changes to `DISABLED` and is resumed if the role changes. This
commit also adds new logs to this class to indicate when jamming is
detected or cleared and/or when any of the parameters get changed.