[meshcop-tlvs] new methods for parsing Channel Mask Entry
This commit adds new methods/classes related to parsing of Channel
Mask Entries in a Channel Mask TLV.

A new class `ChannelMask0Entry` is added for a Channel Mask Entry
with Channel Page 0.

`ChannelMaskTlv::GetFirstEntry()` and`ChannelMaskEntry:GetNext()`
methods are added and can be used to iterate through all the
entries in a Channel Mask TLV.

New method `ChannelMaskTlv::GetMask0Entry()` can be used to search
among all entries to find the one with Channel Page 0 (if one exist).