[CC2538]: Fix MAC address word ordering.
It seems TI (for whatever reason) chose to use a rather unorthadox
mixed-endian representation for the MAC address in the CC2538, where the
most significant 4 octets are given first in little-endian order,
followed by the least significant 4 octets (again in little-endian).

The OpenThread code expects a big-endian representation of the MAC, so
we need to read in and byte-swap each half individually.